requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-phonenumber-field[phonenumbers]
django-rest-swagger